STY - Run through yapf style linter
I ran the script through a standard style linter (yapf) for readability and consistency (this is just the one that I have run on every save). 

I think it looks cleaner and easier to read, and is more consistent, but that’s up to you. No functional changes here, just style rules (e.g. imports were automatically sorted, etc.)

Take it or leave it but I thought it may be helpful!